pyproject.toml
[project]
name = "vae"
version = "0.1.0"
description = "Configuration of the VAE implementation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.0",
    "numpy>=2.2.2",
    "python-dotenv>=1.0.1",
    "torch>=2.5.1",
    "torchvision",
    "tqdm>=4.67.1",
]
